refactor(binders): extract ObservableList and sprite helpers to files
Address review feedback on this PR:

- Move ObservableListViewModelBinderHelper and the sprite CreateSprite helper into their own files (ObservableListViewModelBinderHelper.cs, SpriteBinderHelper.cs) instead of living inside a binder class.
- Call base.OnUnbound() in RendererMaterialColorSwitcherBinder — the override added earlier in this PR was silently dropping the base-class cleanup; also add it to ImageSpriteBinder/Mono to follow the codebase convention.
- Guard CollectionViewModelBinder.OnReplace against an out-of-range index, mirroring the OnAdded guard.
- Add missing trailing newlines on touched files.
