+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Collections/ObservableLists/ViewModel/ObservableListViewModelBinderHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9ccc6d00ecb94c46a0b57f1834c43266
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 0d0d85abe18854107bff963db8208bbc, type: 3}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
