+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Images/Sprite/SpriteBinderHelper.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a6a9648f68d64086b0beb7055d13bdcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {fileID: 2800000, guid: 0d0d85abe18854107bff963db8208bbc, type: 3}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
