+++ Aspid.MVVM/Packages/tech.aspid.mvvm/StarterKit/Unity/Runtime/Binders/Renderers/MaterialsColor/RendererMaterialColorSwitcherBinder.cs
@@ -67,6 +67,7 @@ namespace Aspid.MVVM.StarterKit
         protected override void OnUnbound()
         {
             _materials = null;
+            base.OnUnbound();
         }
     }
 }
